+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added Microsoft SEAL Rust bindings initial version (#2,#3,#4,#5,#7)(@marcosfpr)
 - Added tensor operations for ciphertexts using batching (#8)(@marcosfpr)
+- Added python bindings for sealy (#9)(@marcosfpr)
 
 [unreleased]: https://github.com/marcosfpr/sealy/compare/HEAD...HEAD